cld: Standardize peer ID type to be u_int16_t
Try to make peer ID be 16 bit unsigned type, except where it directly
interacts with the firmware. Although Pronto FW uses just 8 bits for
the peer ID, the txrx layer needs to support at least u_int16_t for
other FWs/HWs, so we standardize on that type, and added 1 assert for
when the u_int16_t is written back into the FW, to ensure overflow
isn't occuring.

CL Id : 2042486

+++ cld/CORE/COMMON/ol_htt_tx_api.h
@@ -599,13 +599,13 @@ htt_tx_desc_display(void *tx_desc);
 #endif
 
 static inline
-void htt_tx_desc_set_peer_id(uint32_t *htt_tx_desc, uint32_t peer_id)
+void htt_tx_desc_set_peer_id(uint32_t *htt_tx_desc, uint16_t peer_id)
 {
     uint32_t *peer_id_field_ptr;
 
     peer_id_field_ptr  = (uint32_t *)(htt_tx_desc + HTT_TX_DESC_PEERID_DESC_PADDR_OFFSET_DWORD);
 
-    *peer_id_field_ptr = (uint32_t)peer_id;
+    *peer_id_field_ptr = peer_id;
 }
